pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "index-all"
version = "0.1.0"
description = "Universal file indexing and summarization engine"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "beautifulsoup4>=4.12,<5",
  "lxml>=5.2,<6",
  "ofxparse>=0.21,<1",
  "openpyxl>=3.1,<4",
  "pandas>=2.2,<3",
  "pypdf>=4.3,<5",
  "python-docx>=1.1,<2",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
